First version of thorougly revised data
This is a first version of a thoroughly revised version of this treebank. 
It was created using the same conversion script as is being used for Dutch LassySmall. As sources, we used the (manually corrected) Alpino treebank annotation for this material as it is available in Groningen. Links to original files have been added. 

Issues: 
* tokenization may differ. In these cases there is no one-to-one mapping from text string to tokens in the annotation. Could be easily fixed if necessary
* some sentences are missing in the Alpino treebanks. In those cases UD 2.0 annotation has been preserved. Alternative: remove these sentences alltogether